pricing_rl_project/sales_agents_project/requirements.txt
# Core
gymnasium==0.29.0
stable-baselines3==2.0.0
torch==2.6.0
numpy==1.24.0
pandas==2.0.0
scikit-learn==1.3.0

# Database
mysql-connector-python==8.1.0
sqlalchemy==2.0.0
redis==4.5.0

# API & Web
fastapi==0.104.0
uvicorn==0.24.0
pydantic==2.0.0
streamlit==1.28.0

# Monitoring
prometheus-client==0.17.0
grafana-api==1.0.3

# Utils
python-dotenv==1.0.0
pyyaml==6.0
matplotlib==3.7.0
seaborn==0.12.0
tensorboard==2.13.0
tqdm==4.66.0

# Testing
pytest==7.4.0
pytest-cov==4.1.0

# Jupyter Notebooks
jupyter==1.0.0
ipykernel==6.25.0
notebook==7.0.0

# Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Logging
python-json-logger==2.0.7

# Additional utilities
click==8.1.7
rich==13.7.0